requirements.txt
black==25.1.0
coverage==7.8.0
django-cors-headers==4.7.0
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-anymail==13.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.5.1
environs[django]==14.2.0
flake8==7.2.0
pillow==11.2.1
psycopg2-binary==2.9.10
